+++ tools/start_compile_docker.sh
@@ -59,7 +59,7 @@
 #  ./tools/build_pip_package.sh ALL_VERSIONS_ALREADY_ASSEMBLED
 #
 # https://hub.docker.com/r/tensorflow/build/tags?page=1
-DOCKER=tensorflow/build:2.13-python3.8
+DOCKER=tensorflow/build:2.14-python3.8
 
 # Current directory
 # Useful if Yggdrasil Decision Forests is available locally in a neighbor